AT Interface: Added a line generator helper and write helper
The generator should make it a lot easier to handle incoming lines by
centralizing the timeout handling -- which is re-entrant, meaning the timeout is
from the start and not for each loop, which is what we want -- and the logging.

Additionally, the writing and checking of the write length is also centralized,
making the checking a little simpler.