AT Interface: Fixed potentially exiting readline() without a newline
Really not finding anything about this, but it's definitely happening where it's
returning with contents but no newline. The subsequent read will return the
remainder of the line.

Maybe it's related to the timeout? Perhaps it's buffering some amount of data,
hitting the timeout before getting a newline, but not putting the data back into
its internal buffering? This might be expected behavior, but I cannot find
anything about it in pyserial's documentation or in the typical forums, so I'm
calling this a bug on their part.