XMODEM: Added better clearing of garbage when starting transmission
Otherwise we might get stuck with some buffered output from something else in
our serial port when we start, and previously we'd only get a single opportunity
to clear that stuff out. We shouldn't end up in the situation where we keep
clearing out our NAK bytes, as the sender shouldn't be sending them so rapidly
that we can't do a single operation between them.