Source: Organized source to have more strict packages
I'm trying to move away from having modules within 'packages' whose classes are
just duplicates of the names.

For example, atError.py containing AtError, which leads to usage that looks
like:

    import nimbelink.cell.modem.atError as atError

    atError.AtError(...)

That's extremely ugly. So, I'm trying to achieve something more like:

    import nimbelink.cell.modem as modem

    modem.AtError(...)

Which, to me, is much more intuitive and easy to understand. The obvious counter
to using __init__.py to perform imports is to ask, "Can't AtError just be
defined in __init__.py?" To which I'll respond, "Ask me when a package has 1,000
lines of multiple classes and helpers."

This also should make it a little easier to come up with names for classes and
such, as you don't have to worry about manually namespacing things. For example:

    from nimbelink.cell.at.result import Result
    from nimbelink.testing.result import Result

To get around the above, you might have been inclined to name the testing result
'TestResult', but with the new approach you can include the context in this
situation:

    import nimbelink.cell.at as at
    import nimbelink.testing as testing

    at.Result(...)
    testing.Result(...)

Each package can import whatever they want to expose -- which is how the concept
of a 'package' should work -- and the help() text can be convinced to show those
imported classes and whatnot using __all__. (Otherwise it would only show the
files it detected and thinks is part of the package.)