Skywire: Added in the sub-modules that the Nano has
With the specifics of the Nano's functionality stripped out, and the function
signatures, interesting interface class members, etc. left in to define how any
modem should implement the sub-modules.

In other words, we're codifying the Nano's design as how any Skywire modem
should be designed, with the hope that it makes implementing additional ones
easier to do while keeping its functionality in line with how the Nano works.