West: Added getting the proper main directory before building
So that others don't have to worry about moving about the file system. We have
all of the information needed from the west environment (for most
configurations, anyway) so we can manage that on our own.

If someone comes across an instance where this gets in the way, they can figure
out the proper workaround ;-)