Command: Added passing args and exceptions to all aborted commands in the stack
Previously, we would only handle KeyboardInterrupt as the 'abort' trigger, and
commands had no context for what happened or what was being run/done. Now we'll
pass the original arguments and any exception to the handler.

Additionally, we would only pass the keyboard interrupt exception to the command
it occurred during, whereas now we'll pass it to each command in the sub-command
chain, giving all sub-command parents a chance to handle cleanup when someone
bad occurs.

This is particularly useful for parent commands that provide resources to
multiple sub-commands, where they might want to also provide the teardown for
their children in addition to the setup.