Nano App: Split up the 'update' stuff into an 'upload' and 'fota'
FOTA is now obvious, and the upload process takes in data rather than a file. We
should also be better about handling the post-DFU reboot stuff. We'll also now
consider the new debug tool if we don't have an AT interface to trigger an
upload with.

The intent of these changes is to allow uploading more than just firmware, such
as keys and partition info.