XMODEM: Changed packet sending to not keep blindly timing out and retrying
This change was not the right way to handle the differences in XMODEM
behavior... We should keep retrying packets if we get a NAK response, per the
protocol. However, if the device simply isn't responding, we shouldn't keep
hammering it with the data over and over. If someone really needs to wait around
forever for a device to respond to a packet -- which the Skywire Nano, for
example, needs during the serial boot loader recovery process -- it's up to the
specific user to change the (new) packet timeout setting.