WSL: Fixed trying to convert path-like things into paths
For instance, if a parameter has the same name as a file or directory in the
current directory, the parameter would get the Windows path to the file or
directory, even though it wasn't necessarily the meaning of the parameter. Files
with a '.' relative path don't need to have their paths updated anyway, so if a
parameter doesn't appear to have a path component, we might as well just skip it
entirely (thus avoiding the potential false file/directory match).