Command: Moved the 'root' command handling into the base Command class
Now that we're starting to deal with potentially having nested 'root' commands,
there's really not a compelling reason to keep these separated anymore. This
also eases the burden a bit on people who are implementing commands, as it
prevents them forgetting to mark their command as potentially 'root' (i.e. use
the separate class).