Wsl: Moved out from 'utils' and into 'command'
This one I'm a little less fond of, but we're hitting a fun circular dependency,
and this is the approach I'm taking to address it.

There are obviously a number of other ways to address this, like changing the
import in 'command' to use a relative path or changing our package design to not
import everything all at once, but I don't like either of those. For the latter,
this would be a world-breaking change to the design, and I've never liked
packages that make you do so much importing. For the former, I want to try and
keep the imports such that sources can be moved around without having to keep
updating all of the imports they perform.