Source: Removed all imports of sub-packages
Turns out this isn't needed for Python to still be happy with using sub-packages
after a single import:

    import nimbelink
    command = nimbelink.command.Command(...)

This allows us to avoid potentially doing a bunch of unneded imports when only
one sub-package is going to get used. We'll still have a little bit of import
overhead with setting up things like modules and commands, but hopefully this
will cut down on them somewhat.

Additionally, this fixes an issue with broken dependencies. For instance, if you
have a pynl package registered that expects another pynl package but the second
package isn't installed, previously you wouldn't have been able to even install
the second package because the module loading done in pynl-base would've
resulted in a ModuleNotFoundError from importing the first package. This should
address those issues entirely.