Module: Changed loading of packages to use a Python-sanctioned loader design
Now we'll register a 'finder' when the top-level 'nimbelink' package is loaded
that will wait for someone to try to import a pynl-like package (i.e.
'nimbelink.<name>'). When we see something that looks like a pynl package being
referenced, *then* we'll look through our list of registered modules and handle
setting up imports for it.

The consequence of this is we lose our previous ability to use lazier importing,
such as:

    import nimbelink
    nimbelink.utils.Xmodem(...)

For 90-95% of our Python source, this won't change anything, since we typically
only use a single package's worth of context. But some import usage will break
with this.

This change also breaks sub-command registration, but that is about to be fixed.

Additionally, this will now fundamentally change how we expect pynl packages to
be designed, since we're no longer actually doing the importing of the packages
but rather appending their install directory to our submodule searching
directories. This will make our pynl packages slightly uglier, since it'll
require essentially an extra, mostly-redundant directory, but it's worth it to
get the importing correct, because it was actually broken before for more
complex situations.